abstract controller method
