+++ src/controller/__init__.py
@@ -2,5 +2,6 @@ from .parser import Parser
 from .controller import Controller
 from .image import Image
 from .lst_parser import LstParser
+from .lst_controller import LstController
 
-__all__ = ['Parser', 'Controller', 'Image', 'LstParser']
+__all__ = ['Parser', 'Controller', 'Image', 'LstParser', 'LstController']
